+++ msl/network/json.py
@@ -74,17 +74,29 @@ def use(value):
 def serialize(obj):
     """Serialize an object as a JSON-formatted string.
 
+    .. versionchanged:: 1.0
+       `obj` can have a callable attribute named `to_json`.
+
     Parameters
     ----------
     obj
-        A JSON-serializable object.
+        A JSON-serializable object. If `obj` is not natively JSON serializable
+        but has a callable attribute named `to_json` then `obj.to_json()` will
+        be called and the returned value will be passed to the `dumps` function
+        of the JSON backend.
 
     Returns
     -------
     :class:`str`
         The JSON-formatted string.
     """
-    out = backend.dumps(obj, **backend.kwargs_dumps)
+    try:
+        out = backend.dumps(obj, **backend.kwargs_dumps)
+    except TypeError as error:
+        try:
+            out = backend.dumps(obj.to_json(), **backend.kwargs_dumps)
+        except AttributeError:
+            raise error from None
     if isinstance(out, bytes):
         return out.decode()
     return out